database_api/requirements.txt
fastapi
uvicorn
requests
mysql-connector-python
pandas
fastapi[all]
statsmodels
pydantic
statsmodels